requirements.txt
# Test framework
pytest>=8.3.4
pytest-bdd>=8.1.0
pytest-playwright>=0.7.1
pytest-asyncio>=0.25.3
pytest-rerunfailures>=15.0
pytest-timeout>=2.4.0
pytest-xdist>=3.6.1
pytest-metadata>=3.1.1
pytest-order>=0.0.0

# Allure report
allure-pytest>=2.13.5
allure-python-commons>=2.13.5

# Browser automation
playwright>=1.49.0

# Environment config
python-dotenv>=1.1.0

# Logging
colorlog>=6.9.0
concurrent-log-handler>=0.9.25

# BDD support
gherkin-official>=29.0.0

# Image comparison
Pillow>=10.0.0
